backend/requirements.txt
fastapi==0.115.12
httpx==0.28.1
numpy==2.2.6
torch==2.11.0
torchaudio==2.11.0
pillow==11.2.1
paddleocr==3.4.1
paddlepaddle
pydantic-settings==2.10.1
python-multipart==0.0.20
soundfile==0.13.1
transformers==5.5.4
uvicorn==0.34.2
pytest==8.3.5
pytest-asyncio==0.26.0
openai>=1.0.0
vieneu>=2.4.3
